1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,6 @@
-#include "function_pointers.h"
 #include <stddef.h>
+
+#include "function_pointers.h"
 
 /**
 * array_iterator - check the code
@@ -10,7 +11,7 @@
 */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i > size; i++)
 		action(array[i]);
